+++ sdk/core/azure-core/inc/azure/core/http/http.hpp
@@ -9,8 +9,8 @@
 #pragma once
 
 #include "azure/core/exception.hpp"
-#include "azure/core/http/body_stream.hpp"
 #include "azure/core/internal/contract.hpp"
+#include "azure/core/io/body_stream.hpp"
 #include "azure/core/nullable.hpp"
 
 #include <algorithm>
@@ -452,7 +452,7 @@ namespace Azure { namespace Core { namespace Http {
     std::map<std::string, std::string> m_headers;
     std::map<std::string, std::string> m_retryHeaders;
 
-    BodyStream* m_bodyStream;
+    Azure::IO::BodyStream* m_bodyStream;
 
     // flag to know where to insert header
     bool m_retryModeEnabled{false};
@@ -469,11 +469,15 @@ namespace Azure { namespace Core { namespace Http {
      *
      * @param httpMethod HTTP method.
      * @param url URL.
-     * @param bodyStream #Azure::Core::Http::BodyStream.
+     * @param bodyStream #Azure::IO::BodyStream.
      * @param downloadViaStream A boolean value indicating whether download should happen via
      * stream.
      */
-    explicit Request(HttpMethod httpMethod, Url url, BodyStream* bodyStream, bool downloadViaStream)
+    explicit Request(
+        HttpMethod httpMethod,
+        Url url,
+        Azure::IO::BodyStream* bodyStream,
+        bool downloadViaStream)
         : m_method(std::move(httpMethod)), m_url(std::move(url)), m_bodyStream(bodyStream),
           m_retryModeEnabled(false), m_isDownloadViaStream(downloadViaStream)
     {
@@ -484,9 +488,9 @@ namespace Azure { namespace Core { namespace Http {
      *
      * @param httpMethod HTTP method.
      * @param url URL.
-     * @param bodyStream #Azure::Core::Http::BodyStream.
+     * @param bodyStream #Azure::IO::BodyStream.
      */
-    explicit Request(HttpMethod httpMethod, Url url, BodyStream* bodyStream)
+    explicit Request(HttpMethod httpMethod, Url url, Azure::IO::BodyStream* bodyStream)
         : Request(httpMethod, std::move(url), bodyStream, false)
     {
     }
@@ -503,7 +507,7 @@ namespace Azure { namespace Core { namespace Http {
         : Request(
             httpMethod,
             std::move(url),
-            NullBodyStream::GetNullBodyStream(),
+            Azure::IO::NullBodyStream::GetNullBodyStream(),
             downloadViaStream)
     {
     }
@@ -515,7 +519,7 @@ namespace Azure { namespace Core { namespace Http {
      * @param url URL.
      */
     explicit Request(HttpMethod httpMethod, Url url)
-        : Request(httpMethod, std::move(url), NullBodyStream::GetNullBodyStream(), false)
+        : Request(httpMethod, std::move(url), Azure::IO::NullBodyStream::GetNullBodyStream(), false)
     {
     }
 
@@ -555,9 +559,9 @@ namespace Azure { namespace Core { namespace Http {
     std::map<std::string, std::string> GetHeaders() const;
 
     /**
-     * @brief Get HTTP body as #Azure::Core::Http::BodyStream.
+     * @brief Get HTTP body as #Azure::IO::BodyStream.
      */
-    BodyStream* GetBodyStream() { return this->m_bodyStream; }
+    Azure::IO::BodyStream* GetBodyStream() { return this->m_bodyStream; }
 
     /**
      * @brief Get the list of headers prior to HTTP body.
@@ -605,7 +609,7 @@ namespace Azure { namespace Core { namespace Http {
     std::string m_reasonPhrase;
     std::map<std::string, std::string> m_headers;
 
-    std::unique_ptr<BodyStream> m_bodyStream;
+    std::unique_ptr<Azure::IO::BodyStream> m_bodyStream;
     std::vector<uint8_t> m_body;
 
     explicit RawResponse(
@@ -613,7 +617,7 @@ namespace Azure { namespace Core { namespace Http {
         int32_t minorVersion,
         HttpStatusCode statusCode,
         std::string const& reasonPhrase,
-        std::unique_ptr<BodyStream> BodyStream)
+        std::unique_ptr<Azure::IO::BodyStream> BodyStream)
         : m_majorVersion(majorVersion), m_minorVersion(minorVersion), m_statusCode(statusCode),
           m_reasonPhrase(reasonPhrase), m_bodyStream(std::move(BodyStream))
     {
@@ -698,11 +702,11 @@ namespace Azure { namespace Core { namespace Http {
     void AddHeader(uint8_t const* const first, uint8_t const* const last);
 
     /**
-     * @brief Set #Azure::Core::Http::BodyStream for this HTTP response.
+     * @brief Set #Azure::IO::BodyStream for this HTTP response.
      *
-     * @param stream #Azure::Core::Http::BodyStream.
+     * @param stream #Azure::IO::BodyStream.
      */
-    void SetBodyStream(std::unique_ptr<BodyStream> stream);
+    void SetBodyStream(std::unique_ptr<Azure::IO::BodyStream> stream);
 
     /**
      * @brief Set HTTP response body for this HTTP response.
@@ -740,9 +744,9 @@ namespace Azure { namespace Core { namespace Http {
     std::map<std::string, std::string> const& GetHeaders() const;
 
     /**
-     * @brief Get HTTP response body as #Azure::Core::Http::BodyStream.
+     * @brief Get HTTP response body as #Azure::IO::BodyStream.
      */
-    std::unique_ptr<BodyStream> GetBodyStream()
+    std::unique_ptr<Azure::IO::BodyStream> GetBodyStream()
     {
       // If m_bodyStream was moved before. nullptr is returned
       return std::move(this->m_bodyStream);