+++ sdk/core/azure-core/test/ut/transport_adapter_base.hpp
@@ -9,7 +9,7 @@
  *
  */
 
-#include <azure/core/http/body_stream.hpp>
+#include <azure/core/io/body_stream.hpp>
 #include <azure/core/http/http.hpp>
 #include <azure/core/internal/http/pipeline.hpp>
 #include <gtest/gtest.h>