+++ sdk/core/perf/test/inc/azure/perf/test/http_client_get_test.hpp
@@ -11,7 +11,7 @@
 
 #include <azure/perf.hpp>
 
-#include <azure/core/http/body_stream.hpp>
+#include <azure/core/io/body_stream.hpp>
 #include <azure/core/http/http.hpp>
 #include <azure/core/http/transport.hpp>
 
@@ -60,7 +60,7 @@ namespace Azure { namespace Perf { namespace Test {
       auto response = Details::HttpClient->Send(ctx, request);
       // Read the body from network
       auto bodyStream = response->GetBodyStream();
-      response->SetBody(Azure::Core::Http::BodyStream::ReadToEnd(ctx, *bodyStream));
+      response->SetBody(Azure::IO::BodyStream::ReadToEnd(ctx, *bodyStream));
     }
 
     /**