+++ sdk/storage/azure-storage-blobs/test/ut/page_blob_client_test.cpp
@@ -38,7 +38,7 @@ namespace Azure { namespace Storage { namespace Test {
     m_blobUploadOptions.HttpHeaders.ContentHash.Value.clear();
     m_pageBlobClient->Create(m_blobContent.size(), m_blobUploadOptions);
     auto pageContent
-        = Azure::Core::Http::MemoryBodyStream(m_blobContent.data(), m_blobContent.size());
+        = Azure::IO::MemoryBodyStream(m_blobContent.data(), m_blobContent.size());
     m_pageBlobClient->UploadPages(0, &pageContent);
     m_blobUploadOptions.HttpHeaders.ContentHash
         = m_pageBlobClient->GetProperties()->HttpHeaders.ContentHash;
@@ -85,7 +85,7 @@ namespace Azure { namespace Storage { namespace Test {
     auto pageBlobClient = Azure::Storage::Blobs::PageBlobClient::CreateFromConnectionString(
         StandardStorageConnectionString(), m_containerName, RandomString());
     pageBlobClient.Create(8_KB, m_blobUploadOptions);
-    auto pageContent = Azure::Core::Http::MemoryBodyStream(blobContent.data(), blobContent.size());
+    auto pageContent = Azure::IO::MemoryBodyStream(blobContent.data(), blobContent.size());
     pageBlobClient.UploadPages(2_KB, &pageContent);
     // |_|_|x|x|  |x|x|_|_|
     blobContent.insert(blobContent.begin(), static_cast<std::size_t>(2_KB), '\x00');
@@ -120,7 +120,7 @@ namespace Azure { namespace Storage { namespace Test {
     auto snapshot = pageBlobClient.CreateSnapshot()->Snapshot;
     // |_|_|_|x|  |x|x|_|_| This is what's in snapshot
     blobContent.resize(static_cast<std::size_t>(1_KB));
-    auto pageClient = Azure::Core::Http::MemoryBodyStream(blobContent.data(), blobContent.size());
+    auto pageClient = Azure::IO::MemoryBodyStream(blobContent.data(), blobContent.size());
     pageBlobClient.UploadPages(0, &pageClient);
     pageBlobClient.ClearPages({3_KB, 1_KB});
     // |x|_|_|_|  |x|x|_|_|
@@ -247,7 +247,7 @@ namespace Azure { namespace Storage { namespace Test {
     auto pageBlobClient = Azure::Storage::Blobs::PageBlobClient::CreateFromConnectionString(
         StandardStorageConnectionString(), m_containerName, RandomString());
     pageBlobClient.Create(blobContent.size(), m_blobUploadOptions);
-    auto pageContent = Azure::Core::Http::MemoryBodyStream(blobContent.data(), blobContent.size());
+    auto pageContent = Azure::IO::MemoryBodyStream(blobContent.data(), blobContent.size());
 
     Blobs::UploadPageBlobPagesOptions options;
     ContentHash hash;
@@ -275,7 +275,7 @@ namespace Azure { namespace Storage { namespace Test {
     auto pageBlobClient = Azure::Storage::Blobs::PageBlobClient::CreateFromConnectionString(
         StandardStorageConnectionString(), m_containerName, RandomString());
     pageBlobClient.Create(blobContent.size(), m_blobUploadOptions);
-    auto pageContent = Azure::Core::Http::MemoryBodyStream(blobContent.data(), blobContent.size());
+    auto pageContent = Azure::IO::MemoryBodyStream(blobContent.data(), blobContent.size());
 
     Blobs::UploadPageBlobPagesOptions options;
     ContentHash hash;
@@ -307,7 +307,7 @@ namespace Azure { namespace Storage { namespace Test {
     }
 
     auto blobContent
-        = Azure::Core::Http::MemoryBodyStream(m_blobContent.data(), m_blobContent.size());
+        = Azure::IO::MemoryBodyStream(m_blobContent.data(), m_blobContent.size());
     blobClient.UploadPages(0, &blobContent);
     {
       auto response = blobClient.CreateIfNotExists(m_blobContent.size());
@@ -315,7 +315,7 @@ namespace Azure { namespace Storage { namespace Test {
     }
     auto downloadStream = std::move(blobClient.Download()->BodyStream);
     EXPECT_EQ(
-        Azure::Core::Http::BodyStream::ReadToEnd(Azure::Core::Context(), *downloadStream),
+        Azure::IO::BodyStream::ReadToEnd(Azure::Core::Context(), *downloadStream),
         m_blobContent);
   }