+++ sdk/storage/azure-storage-files-datalake/sample/datalake_getting_started.cpp
@@ -64,12 +64,12 @@ void DataLakeGettingStarted()
     std::vector<uint8_t> buffer(str1.begin(), str1.end());
 
     // One way of passing in the buffer, note that the buffer is not copied.
-    auto bufferStream = Azure::Core::Http::MemoryBodyStream(buffer);
+    auto bufferStream = Azure::IO::MemoryBodyStream(buffer);
 
     fileClient.Append(&bufferStream, 0 /* Offset of the position to be appended.*/);
 
     // Another way of passing in the buffer, note that buffer is also not copied.
-    bufferStream = Azure::Core::Http::MemoryBodyStream(
+    bufferStream = Azure::IO::MemoryBodyStream(
         reinterpret_cast<const uint8_t*>(str2.data()), str2.size());
 
     fileClient.Append(&bufferStream, str1.size());
@@ -81,7 +81,7 @@ void DataLakeGettingStarted()
     auto result = fileClient.Download();
     Azure::Core::Context context;
     std::vector<uint8_t> downloaded
-        = Azure::Core::Http::BodyStream::ReadToEnd(context, *(result->Body));
+        = Azure::IO::BodyStream::ReadToEnd(context, *(result->Body));
     // downloaded contains your downloaded data.
     std::cout << "Downloaded data was:\n" + std::string(downloaded.begin(), downloaded.end())
               << std::endl;